backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.6.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytesseract==0.3.10
python-multipart==0.0.6
python-docx==1.1.0
pypdf2==3.0.1
pillow==10.1.0
email-validator==2.1.1
minio==7.2.7
reportlab==4.2.2